Moved ProductCache and OrderCache to their respective folder
